% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu1col.R
\name{elmSet}
\alias{elmSet}
\title{Function to set element numbers of the intermediate values at time i}
\usage{
elmSet(n_time, n_grp, t)
}
\arguments{
\item{n_time}{Length of the time series}

\item{n_grp}{Number of areas}

\item{i}{Time point i}
}
\description{
Function to set element numbers of the intermediate values at time i
}
