% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendepend.R
\name{genDepnd}
\alias{genDepnd}
\title{Function to generate dependent variable}
\usage{
genDepnd(x, beta, n_grp, rho, omega)
}
\arguments{
\item{x}{Independent variable. Vector or matrix.}

\item{beta}{Coefficients of x.}

\item{rho}{Parameter of serial correlation.}

\item{omega}{Parameter of synchronization.}
}
\description{
genDepnd function is one of the simulation functions of the package timsync.
genDepnd generates the value of dependent variable from the independent x and
its/their coefficient beta.
}
