% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendepend.R
\name{genDepndIni}
\alias{genDepndIni}
\title{Function to generate Initial value}
\usage{
genDepndIni(x, res_elm, beta, Sig, omega)
}
\arguments{
\item{rel_elm}{Position of the focul element in the data x and y.}
}
\description{
Function to generate Initial value
}
