% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendepend.R
\name{genDepndSeq}
\alias{genDepndSeq}
\title{Function to generate dependent variable sequentially}
\usage{
genDepndSeq(x, y_old, res_elm, beta, Sig, rho, omega)
}
\arguments{
\item{x}{independent variables}

\item{y_old}{previous values one lag before the generating time point}
}
\description{
Function to generate dependent variable sequentially
}
