% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timix.R
\name{gridprm}
\alias{gridprm}
\title{Grid search of rho and omega}
\usage{
gridprm(rho_vc, omega_vc, dat, y, f, r, grp)
}
\arguments{
\item{rho_ve}{vector of the parameter rho}

\item{omega_ve}{vector of the parameter omega}
}
\description{
This function conduct the grid search for selecting the best parameter set of rho and omega
}
