% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu1col.R
\name{makeGrp}
\alias{makeGrp}
\title{Function to  make group}
\usage{
makeGrp(n_time, n_grp)
}
\arguments{
\item{n_time}{Length of the time series}

\item{n_grp}{Number of areas}
}
\description{
Function to  make group
}
