% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendat.R
\name{makeOmega}
\alias{makeOmega}
\title{Function to generate Omega matrix}
\usage{
makeOmega(n_time, n_grp, rho, omega)
}
\arguments{
\item{n_time}{Length of the time series}

\item{n_grp}{Number of groups}

\item{rho}{Parameter of serial correlation}

\item{omega}{Parameter of synchronization}
}
\description{
Function to generate Omega matrix
}
