% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu1col.R
\name{makeRandNum}
\alias{makeRandNum}
\title{Make random numbers with synchronization}
\usage{
makeRandNum(rndVar, mean, sd, Sig)
}
\arguments{
\item{mean}{Mean of the random numbers}

\item{Sig}{Variance-Covariance matrix to prescribe synchronization}
}
\description{
Make random numbers with synchronization
}
