% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu1col.R
\name{makeSyncMat}
\alias{makeSyncMat}
\title{Make the variance-covariance matrix to prescribe synchronization}
\usage{
makeSyncMat(omega, n_grp)
}
\arguments{
\item{omega}{Parameter of synchronization}

\item{n_grp}{Number of areas}
}
\description{
Make the variance-covariance matrix to prescribe synchronization
}
