% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmLogLik.R
\name{mmLogLik}
\alias{mmLogLik}
\title{Function to calculate maximum likelihood}
\usage{
mmLogLik(rho, omega, dat, y, f, r, grp, w, b, Ve, Vu, method)
}
\arguments{
\item{rho}{parameter for serial correlation}

\item{omega}{parameter for synchronization}

\item{dat}{data for the analysis}

\item{y}{column number of the dependent variable}

\item{f}{column number of the fixed effects}

\item{r}{column number of the random effects}

\item{grp}{column number of the group}

\item{w}{weight vector for the random effects variance}

\item{Ve}{Variance of the residuals}

\item{Vu}{Variance of the random effects}

\item{method}{"ML" or "REML"}
}
\description{
mmLogLik function is to calculate the maximum likelihood values
under the estimated parameters.
The appropriate usage is to calculate the ML under the REML parameters
estiamted by the timixCov or timix function.
}
\references{
Hamazaki, K., & Iwata, H. (2020).
RAINBOW: Haplotype-based genome-wide association study
 using a novel SNP-set method. PLoS computational biology, 16(2), e1007663.
}
