% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu1col.R
\name{simuOneCol}
\alias{simuOneCol}
\title{Function to generate data of one column}
\usage{
simuOneCol(n_time, n_grp, rho, omega, rndVar = F, mean = 0, sd = 0)
}
\arguments{
\item{n_time}{Length of the time series}

\item{n_grp}{Number of groups}

\item{rho}{Parameter of serial correlation}

\item{omega}{Parameter of synchronization}

\item{rndVar}{Set random numbers for the each group mean.
Default is FALSE}

\item{mean}{mean value (hyper parameter) in the distribution of the group mean.
Default is 0}

\item{sd}{sd value (hyper parameter) in the distribution of the group mean.
Default is 0}
}
\description{
Function to generate data of one column
}
\examples{
n_time <- 100
n_grp <- 5

y <- simuOneCol(n_time, n_area, rho = 0.5, omega = 0.5,
 rndVar = T, mean = 0, sd = 1)
x0 <- rep(1, n_time * n_grp) # Intercept
x1 <- simuOneCol(n_time, n_grp, rho = 0.5, omega = 1)
x2 <- simuOneCol(n_time, n_grp, rho = 0.5, omega = 0)
area <- makeGrp(n_time, n_grp)

dat <- data.frame(y, x0, x1, x2, area)

}
