% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfixed.R
\name{testFixed}
\alias{testFixed}
\title{F-test for the fixed effects}
\usage{
testFixed(X, y, beta, sigma_sq, V, n, p)
}
\arguments{
\item{X}{design matrix for the fixed effects.}

\item{y}{dependent variable.}

\item{beta}{a vector of the estimated coefficients.}

\item{sigma_sq}{the estimated sigma square, a component of the denominator of the F value.}

\item{V}{the variance-covariance matrix of the dependent variavble.}

\item{n}{the number of data}

\item{p}{the number of fixed effects}
}
\description{
F-test for the fixed effects
}
