% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timix.R
\name{timix}
\alias{timix}
\title{Mixed model for the synchronized time sereis with given covariance parameters.}
\usage{
timix(rho, omega, dat, y, f, r, grp)
}
\arguments{
\item{rho}{parameter for serial correlation}

\item{omega}{parameter for synchronization}

\item{dat}{data for the analysis}

\item{y}{column number of the dependent variable}

\item{f}{column number of the fixed effects}

\item{r}{column number of the random effects}

\item{grp}{column number of the group}
}
\description{
timix function is a mix model function to solve synchronized time sereis.
The effects of synchronization and serial correlation are treated as the random effects.
The variance-covariance matrix of random effects are prescribed by the parameters rho and omega.
}
\examples{
n_time <- 50
n_grp <- 5
y <- simuOneCol(n_time, n_grp, rho = 0.5, omega = 0.5, mean = 0)
x0 <- rep(1, n_time * n_grp) # Intercept
x1 <- simuOneCol(n_time, n_grp, rho = 0.5, omega = 1, mean = 0)
x2 <- simuOneCol(n_time, n_grp, rho = 0.5, omega = 0, mean = 0)
area <- makeGrp(n_time, n_grp)
dat <- data.frame(y, x0, x1, x2, area)
timix(rho = 0.5, omega = 0.5, dat, y = 1, f = 2:4, r = 2:4, grp = 5)
}
