% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt.R
\name{timixCovRho}
\alias{timixCovRho}
\title{Mixed model for the synchronized time sereis by optimization, when omega = 0}
\usage{
timixCovRho(dat, y, f, r, grp, grid = T, trace = F)
}
\arguments{
\item{dat}{data for the analysis}

\item{y}{column number of the dependent variable}

\item{f}{column number of the fixed effects}

\item{r}{column number of the random effects}

\item{grp}{column number of the group}
}
\description{
timixCovRho function is a mixed model function to solve synchronized time sereis when omega = 0.
The effects of serial correlation is treated as the random effects.
The variance-covariance matrix of random effects are prescribed by the parameter rho.
timixCovRho function estimate the parameter through oprimization.
}
\examples{
n_time <- 50
n_grp <- 5
y <- simuOneCol(n_time, n_grp, rho = 0.5, omega = 0.5, mean = 0)
x0 <- rep(1, n_time * n_grp) # Intercept
x1 <- simuOneCol(n_time, n_grp, rho = 0.5, omega = 1, mean = 0)
x2 <- simuOneCol(n_time, n_grp, rho = 0.5, omega = 0, mean = 0)
area <- makeGrp(n_time, n_grp)
dat <- data.frame(y, x0, x1, x2, area)
res <- timixCovRho(dat, y = 1, f = 2:4, r = 2:4, grp = 5)

}
