% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timix.R
\name{timixRev}
\alias{timixRev}
\title{obtain negative log likelihood of timix function for optimization process}
\usage{
timixRev(prm, dat, y, f, r, grp)
}
\arguments{
\item{prm}{prm[1] is rho and prm[2] is omega.
This notification is for oprimization process}
}
\description{
obtain negative log likelihood of timix function for optimization process
}
