% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt.R
\name{timixRevRho}
\alias{timixRevRho}
\title{obtain negative log likelihood of timix function for optimization process}
\usage{
timixRevRho(prm, dat, y, f, r, grp)
}
\arguments{
\item{prm}{prm[1] is rho, while omega is set to be 0.
This notification is for oprimization process}
}
\description{
obtain negative log likelihood of timix function for optimization process
}
