% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timsyncEMMREML.R
\name{tsEMMREMLMultiKernel}
\alias{tsEMMREMLMultiKernel}
\title{Fit mixed model}
\usage{
tsEMMREMLMultiKernel(
  y,
  X,
  Zlist,
  Klist,
  varbetahat = TRUE,
  varuhat = TRUE,
  PEVuhat = FALSE,
  test = TRUE
)
}
\arguments{
\item{y}{n x 1 vector of explained variable}

\item{X}{n x q design matrix of explanatory variables}

\item{Zlist}{list of random effects design matrices}

\item{Klist}{list of variance-covariance matrices of random effects}
}
\description{
This function is modified from the emmremlMultiKernel function
in the EMMREML package (Akdemir & Godfrey, 2015).
}
\references{
Akdemir, D., & Okeke, U. G. (2015).
EMMREML: Fitting mixed models with known covariance structures.
R package version, 3.1.
}
