% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu1col.R
\name{simuOneCol}
\alias{simuOneCol}
\title{Function to generate data of one column}
\usage{
simuOneCol(n_time, n_grp, rho, omega, mean = 0)
}
\arguments{
\item{n_time}{Length of the time series}

\item{n_grp}{Number of groups}

\item{rho}{Parameter of serial correlation}

\item{omega}{Parameter of synchronization}
}
\description{
Function to generate data of one column
}
\examples{
n_time <- 100
n_grp <- 5

y <- simuOneCol(n_time, n_area, rho = 0.5, omega = 0.5, mean = 0)
x0 <- rep(1, n_time * n_area) # Intercept
x1 <- simuOneCol(n_time, n_area, rho = 0.5, omega = 1, mean = 0)
x2 <- simuOneCol(n_time, n_area, rho = 0.5, omega = 0, mean = 0)
area <- makeGrp(n_time, n_area)

dat <- data.frame(y, x0, x1, x2, area)

}
